requirements.txt
neo4j==4.4.1
requests==2.26.0
scipy==1.7.3
spacy==3.2.0
stanford_openie==1.2.0
nltk==3.6.6
six==1.16.0